fix mypy and memory measurements errors in HTs
Former-commit-id: ccd7aadf5996a3d4031a2d9676b05f8b7a7e789d

+++ river/tree/hoeffding_tree_regressor.py
@@ -118,7 +118,7 @@ class HoeffdingTreeRegressor(HoeffdingTree, base.Regressor):
         splitter: Splitter = None,
         min_samples_split: int = 5,
         binary_split: bool = False,
-        max_size: int = 500,
+        max_size: float = 500.0,
         memory_estimate_period: int = 1000000,
         stop_mem_management: bool = False,
         remove_poor_attrs: bool = False,
